+++ src/pages/payment/PaymentPage.tsx
@@ -1,7 +1,7 @@
 import * as S from './PaymentPage.styled';
 import Header from '@shared/components/Header/Header';
-import { usePageNavigation } from '@/shared/hooks/usePageNavigation';
-import BottomConfirmButton from '@/shared/components/BottomConfirmButton/BottomConfirmButton';
+import { usePageNavigation } from '@app/hooks/usePageNavigation';
+import BottomConfirmButton from '@shared/components/BottomConfirmButton/BottomConfirmButton';
 
 export default function PaymentPage() {
   const { getPaymentSuccessState, navigateToCart } = usePageNavigation();
